captcha_module/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
kiwisolver==1.2.0
lxml==4.5.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.17.2
scipy==1.4.1
selenium==3.141.0
six==1.14.0
SpeechRecognition==3.8.1
tifffile==2020.5.11
urllib3==1.25.9
validators==0.15.0
